+++ x/msgfee/init_test.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/iov-one/weave"
 	"github.com/iov-one/weave/coin"
+	"github.com/iov-one/weave/migration"
 	"github.com/iov-one/weave/store"
 )
 
@@ -30,6 +31,7 @@ func TestGenesis(t *testing.T) {
 	}
 
 	db := store.MemStore()
+	migration.MustInitPkg(db, "msgfee")
 	var ini Initializer
 	if err := ini.FromGenesis(opts, db); err != nil {
 		t.Fatalf("cannot load genesis: %s", err)
@@ -69,6 +71,7 @@ func TestGenesisWithInvalidFee(t *testing.T) {
 			}
 
 			db := store.MemStore()
+			migration.MustInitPkg(db, "msgfee")
 			var ini Initializer
 			if err := ini.FromGenesis(opts, db); err == nil {
 				t.Fatal("no error")