+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Breaking changes
 - x/namecoin is using schema versioned model and messages
 - x/sigs is using schema versioned model and messages
 - x/multisig is using schema versioned model and messages
+- x/msgfee is using schema versioned model and messages
 
 
 ## 0.13.0